Create separate test helpers file
