+++ test/pattern_properties.test.js
