+++ test/properties.test.js
