+++ test/readonly_writeonly.test.js
